+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Bug Fixes
 
 1. Fixed stalled Zcash signing when response QR generation fails
+2. Restricted Zcash shielded signing to the selected account
 
 
 ## 2.5.0(2026-6-29)